duplicated test on mnemonic and encrypted
